refactor: ContributorItem 함수 제거
